+++ lib/types.d.ts
@@ -1,6 +1,6 @@
 export declare class StstConfig {
-    port: number;
-    root: string;
+    port?: number;
+    root?: string;
     constructor(port: number, root: string);
 }
 export type Literal = {