+++ src/types.ts
@@ -1,6 +1,6 @@
 export class StstConfig {
-    port: number
-    root: string
+    port?: number
+    root?: string
     constructor (port: number, root: string) {
         this.port = port
         this.root = root