+++ lib/stream-statics.d.ts
@@ -1,4 +1,5 @@
 import * as http from 'http';
-import { StstConfig } from './types';
-export declare const defaultConfig: StstConfig;
-export declare const startServer: (inputConfig: StstConfig) => http.Server<typeof http.IncomingMessage, typeof http.ServerResponse>;
+export declare const startServer: (inputConfig: {
+    root?: string;
+    port?: number;
+}) => http.Server<typeof http.IncomingMessage, typeof http.ServerResponse>;