+++ lib/types.d.ts
@@ -1,9 +1,0 @@
-export declare class StstConfig {
-    port?: number;
-    root?: string;
-    constructor(port: number, root: string);
-}
-export type Literal = {
-    [key: string]: string;
-};
-export type StringOrFalse = string | false;